refactor: reorganise Hero component structure and dependencies
- Refactored Hero component file structure, moving it to `src/components/Hero/Hero.tsx`.
- Updated imports in `src/app/page.tsx` to reflect the new Hero component path.
- Replaced `HeroSocialLinks` with `SocialLinks` and updated import path.
- Removed related Hero tests (`src/tests/hero.test.tsx`) since the component structure changed.
- Added/tested new dev dependencies such as `msw`, `jest-axe`, and updates to `@testing-library` packages.
- Updated `package-lock.json` with new and updated dependency versions.
